Generate an order dependant widget chain name and hash it to fix issues with groups having similar id.

+++ src/sgui/Internals/GuiGroup.h
@@ -14,7 +14,7 @@ struct GroupData
   bool isActive = false;
   bool horizontal = false;
   bool hasMenuBar = false;
-  uint32_t groupId = 0u;
+  size_t groupId = 0u;
   uint32_t clippingLayer = 0u;
   uint32_t plotterLayer = 0u;
   uint32_t menuItemCount = 0u;
